requirements.txt
Django>=4.2.0,<4.3.0
djangorestframework>=3.14.0,<4.0.0
django-filter>=23.2.0,<24.0.0
pytz>=2023.3,<2024.0
django-cors-headers>=3.14.0,<4.0.0
djoser>=2.2.0,<3.0.0
djangorestframework_simplejwt>=5.2.2,<6.0.0
django-debug-toolbar>=4.0.0,<5.0.0
channels>=4.0.0,<5.0.0
daphne>=4.0,<5
twisted[http2,tls]
channels_redis>=4.1.0,<5.0.0
pyyaml>=6.0.0,<7.0.0
uritemplate>=4.1.1,<5.0.0
pytest>=7.3.0,<8.0.0
